Add simple section plane
The section plane works a bit like the section plane of the arch module.
It cuts all included objects and creates a SVG from the section cut.

But it does only a minimal part of the original SectionPlane in the
arch module. It only displays the sections not all faces, uses hatch
patterns with colors, and produces a full svg with the content centered
and scaled.