Rendering of noncutting planes
When we render planes, that do not actually cut the model, nothing was
rendered. This commit fixes it by creating a dummy plane in this case that
only creates secondary faces.

In Addition we fix sorting. Previously the objects where sorted. But we
need to sort faces after cutting. Oterhwise when a large object has
multiple faces cut, it can happen that they are in the wrong order.