Fix lost sorting after duplicate face removal
Before, we replaced the previous face with the new one. But this reversed
orientation, because the face that comes later in the original list,
will be inserted before faces, that where originally before the actual face.

By simply removing the previous face, and adding the new one to the end,
we preserve the original order in the list.